Drop call to local_build_parameters_list() and leave to main_function instead.
R CMD CHECK was producing a NOTE saying that there was no definition of
local_build_parameters_list().  It was only being called here if neither tzar
nor tzar emulation were being done and in that case, it would be odd to be
calling this tzar emulation library at all.  So, I've decoupled that from
this library.  If the mainline needs a parameters list and isn't using tzar
or tzar emulation, it will have to handle that itself - which it would have
had to do anyway because it would have had to provide the local_build...()
routine anyway.