Make emulation flag first arg instead of last because it changes the most.
Before this change, the emulatingTzar flag was the last argument to
run_mainline_under_tzar_or_tzar_emulation(), even though it was the value that
would change most often.  That meant that all other values in the list would
always have to be explicitly set and they will nearly always want to default instead.