Make it possible to copy a model.R into R directory when doing pkg development
tzar expects there to be a model.R file in with a project's R code so that it
can call the top level of the project application code.  However, when building
a package, that model.R code will be automatically executed by the build and
it will contain code that should not be run at build time.
Changes in this commit make it possible to choose whether to have model.R in the
R directory (e.g., when doing normal non-package coding) or to create model.R
each time tzar is run and get rid of it after tzar finishes the run.