Added code to trap emulation crash and save console output file
Whenever tzar emulation was crashing, it would not close the console
sink file and the important information about what went wrong would be
lost, particularly information about random seed settings.
This new code wraps the main call in a tryCatch() and makes sure that
the console output file is saved when there is a fatal error in a run.

Tested by adding these 3 lines at the start of bdgpxupaper::xu_paper_main()

cat ("\n\nAbout to crash tzar to test saving the console sink.\n")
stop ("Test crash of tzar and bdpgxupaper")
cat ("\n\nImmediately after crash tzar to test saving the console sink.  SHOULD NOT REACH THIS STATEMENT.\n")

The corresponding commit in bdpgxupaper for the test was:
- 0a297248
- Special version to test new tzar trapping fatal errors & save console output
- 2017-12-13 08:30 (strangely not Melbourne time...)